requirements.txt
flask
fpdf
python-docx
PyMuPDF
gunicorn